+++ compiler/src/main/java/com/squareup/anvil/compiler/Utils.kt
@@ -110,9 +110,9 @@ public const val OPTION_GENERATE_SHIMS: String = "anvil.ksp.generateShims"
  * KSP option to specify custom extra contributing annotations. Useful for situations where you
  * can't or don't want to implement an `AnvilKspExtension` implementation to provide them.
  *
- * Value should be a comma-separated list of fully qualified canonical class names.
+ * Value should be a colon-delimited list of fully qualified canonical class names.
  */
-public const val OPTION_EXTRA_CONTRIBUTING_ANNOTATIONS: String = "anvil.ksp.extraContributingAnnotations"
+public const val OPTION_EXTRA_CONTRIBUTING_ANNOTATIONS: String = "anvil-ksp-extraContributingAnnotations"
 
 /**
  * Returns the single element matching the given [predicate], or `null` if element was not found.